0.15.6 — the develop pipeline knows what its numbers mean
Culling and grading are untouched.

Exposure inverts the actual composite it sits on (decoder gamma → camera
calibration → standard profile curve) instead of assuming sRGB. Against
LibRaw computing one more stop from linear sensor data, pixel error at
+2.0 EV drops from 11.65/8.78/13.96 to 1.61/0.19/0.98 on A6700, S5M2X and
R6 Mark III. Matching only the decoder gamma is worse — the profile curve
dominates. JPEG and HEIF originals are inverted with sRGB, since they never
went through demosaic or a profile. Match camera JPEG fits in the same
space: hand it an image the engine brightened by +1 EV and the fit recovers
+1 EV.

Automatic lens vignetting had the same wrong curve. Flatness error on a
known vignette goes from 1.57 levels to 0.00, and corner residuals that ran
−0.36 to +0.21 stops depending on scene brightness are gone. The earlier
check missed it by applying and removing the vignette through the same
sRGB curve — a round trip reads as correct even when the curve is wrong.

Manual vignetting is now linear in stops. The same number used to mean
+0.43 stops on a dark corner and +1.16 on a bright one; setting it by eye
on the dark corners blew out the bright ones (clipped pixels 0.08% → 1.68%
at 50, now 0.26%). Saved values for that slider mean something different.

Colour management: Adobe RGB JPEG and HEIF originals open correctly with
per-file detection; export can target sRGB or Adobe RGB with the profile
embedded; TIFF and PNG can be written at 16-bit, with 8-bit output
unchanged to the level; HSL and fringe removal work in float instead of
8-bit HSV.

Also: Panasonic RW2 can be demosaiced for analysis instead of using its
10%-size embedded preview, highlight recovery is available as a develop
option, and HIF is read at 10-bit for develop.

Suite: 1820 passed, 6 skipped.

+++ arw_selector/core/develop/exif_strip.py
@@ -219,7 +219,10 @@ def apply_exif_strip(
         line_color = (60, 60, 64) if settings.dark_background else (200, 200, 203)
         cv2.line(strip, (0, 0), (width, 0), line_color, 1)
 
-        return np.vstack([image, strip])
+        # 띠는 8비트로 그리지만(글자·배경이라 계조가 필요 없습니다) 사진의
+        # dtype에 맞춰 붙입니다. 16비트로 내보낼 때 그대로 vstack하면 띠
+        # 값(0~255)이 0~65535 눈금에서 거의 검게 나옵니다.
+        return np.vstack([image, strip.astype(image.dtype)])
     except Exception as exc:  # noqa: BLE001 - 띠 실패로 내보내기를 막지 않습니다
         log.warning("EXIF 띠 생성 실패: %s", exc)
         return image
